requirements.txt
Cerberus==1.3.4
click==8.0.1
colorama==0.4.4
Flask==2.0.1
greenlet==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
psycopg2==2.9.1
pydantic==1.8.2
python-dotenv==0.19.0
SQLAlchemy==1.4.23
sqlalchemy2-stubs==0.0.2a15
sqlmodel==0.0.4
typing-extensions==3.10.0.2
Werkzeug==2.0.1
